Added convolution test script
This script is meant to be an example to show two things. First that we
can verify if an IR is correct by convolving a recovered IR with the
MLS and comparing the output to the original recording. Second, the
IR's currently being recovered from the data are not correct.