Updated (working) OverlapSaveDemo
I've corrected the problems with this program, it now gives correct
answers.  The only thing I've managed to do that causes it to not work
properly is set the order (N) only one longer than the number of samples
in "h" the impulse response (so if "h" is six samples long, setting N=7
causes problems but N=8 and higher doesn't).  The program actually still
gives the proper answer, it's just missing the last two digits of the
convolution.  Everything else is there and accurate.