FFT, IFFT, and supporting Files
Here are [what I hope to be] the final versions of the FFT and IFFT
functions.  Included also are fftstuff (.c and .h files); the .c has
some functions in it that get called by FFT and IFFT, while the .h file
is of course the corresponding header.

Full descriptions of FFT and IFFT are in the comments inside the
functions themselves (FFT has the most detailed desc.)